init example; renamed nx3d.main to nx3d.core
