+++ src/nx3d/__init__.py
@@ -1,1 +1,1 @@
-from .main import Nx3D, demo, plot
+from .core import Nx3D, demo, plot
