+++ src/nx3d/examples.py
@@ -1,0 +1,1 @@
+from nx3d.core import Nx3D
